requirements.txt
numpy==1.16.4
matplotlib==3.0.3
Flask==1.1.1
keras==2.3.1
Pillow==7.0.0
